requirements.txt
streamlit>=1.32.0
openai-whisper>=20231117
anthropic>=0.25.0
edge-tts>=6.1.9

soundfile>=0.12.1
numpy>=1.24.0
python-dotenv>=1.0.0

ffmpeg-python>=0.2.0
groq